addressed review comments #3
