+++ Tasks/AzureAppServiceManage/azureappservicemanage.ts
@@ -1,7 +1,7 @@
 import tl = require('vsts-task-lib/task');
 import Q = require('q');
 import path = require('path');
-import { AzureRMEndpoint } from 'azure-arm-rest/azure-arm-common';
+import { AzureRMEndpoint } from 'azure-arm-rest/azure-arm-endpoint';
 import { AzureEndpoint } from 'azure-arm-rest/azureModels';
 import {AzureAppService  } from 'azure-arm-rest/azure-arm-app-service';
 // import { AzureAppService } from 'azure-arm-rest/AzureAppService';
